+++ tests/integration/test_coinbase.py
@@ -2,19 +2,15 @@ from tests.test_helper import *
 
 from braintree.test.nonces import Nonces
 from braintree.exceptions.not_found_error import NotFoundError
+from braintree.error_codes import ErrorCodes
 
 class TestCoinbase(unittest.TestCase):
 
-    def _assert_valid_coinbase_account(self, account):
-        for attr in ["user_name", "user_email", "user_id"]:
-            self.assertIsNotNone(getattr(account, attr))
-
     def test_customer(self):
         result = Customer.create({"payment_method_nonce": Nonces.Coinbase})
-        customer = Customer.find(result.customer.id)
-        account = customer.coinbase_accounts[0]
-        self.assertIsNotNone(account)
-        self._assert_valid_coinbase_account(account)
+
+        self.assertFalse(result.is_success)
+        self.assertEquals(ErrorCodes.PaymentMethod.PaymentMethodNoLongerSupported, result.errors.for_object("coinbase_account").on("base")[0].code)
 
     def test_vault(self):
         result = Customer.create()
@@ -23,14 +19,11 @@ class TestCoinbase(unittest.TestCase):
             "payment_method_nonce": Nonces.Coinbase
         })
 
-        account = result.payment_method
-        self._assert_valid_coinbase_account(account)
-
-        PaymentMethod.delete(account.token)
-
-        self.assertRaises(braintree.exceptions.NotFoundError, PaymentMethod.find, account.token)
+        self.assertFalse(result.is_success)
+        self.assertEquals(ErrorCodes.PaymentMethod.PaymentMethodNoLongerSupported, result.errors.for_object("coinbase_account").on("base")[0].code)
 
     def test_transaction(self):
         result = Transaction.sale({"payment_method_nonce": Nonces.Coinbase, "amount": "1.00"})
-        account = result.transaction.coinbase_details
-        self._assert_valid_coinbase_account(account)
+
+        self.assertFalse(result.is_success)
+        self.assertEquals(ErrorCodes.PaymentMethod.PaymentMethodNoLongerSupported, result.errors.for_object("transaction").on("base")[0].code)